requirements.txt
langgraph
langchain-core
langchain-google-genai
langsmith
python-dotenv
ipython
notebook
requests
BeautifulSoup